refactor : ArtworkResponse 패키지 이동
